Extract libp2p module from network module
