+++ bin/full-node/src/sync_service.rs
@@ -31,7 +31,7 @@ use futures::{
     prelude::*,
 };
 use std::{collections::BTreeMap, sync::Arc, time::SystemTime};
-use substrate_lite::{chain::sync::optimistic, database::full_sled, network};
+use substrate_lite::{chain::sync::optimistic, database::full_sled, libp2p, network};
 use tracing::Instrument as _;
 
 /// Configuration for a [`SyncService`].
@@ -48,7 +48,7 @@ pub struct Config {
 pub enum Event {
     BlocksRequest {
         id: BlocksRequestId,
-        target: network::PeerId,
+        target: libp2p::PeerId,
         request: network::protocol::BlocksRequestConfig,
     },
 }
@@ -141,7 +141,7 @@ impl SyncService {
 
     /// Registers a new source for blocks.
     #[tracing::instrument(skip(self))]
-    pub async fn add_source(&self, peer_id: network::PeerId, best_block_number: u64) {
+    pub async fn add_source(&self, peer_id: libp2p::PeerId, best_block_number: u64) {
         self.to_background
             .lock()
             .await
@@ -152,7 +152,7 @@ impl SyncService {
 
     /// Removes a source of blocks.
     #[tracing::instrument(skip(self))]
-    pub async fn remove_source(&self, peer_id: network::PeerId) {
+    pub async fn remove_source(&self, peer_id: libp2p::PeerId) {
         self.to_background
             .lock()
             .await
@@ -165,7 +165,7 @@ impl SyncService {
     ///
     /// Has no effect if the previously-known best block is lower than the new one.
     #[tracing::instrument(skip(self))]
-    pub async fn raise_source_best_block(&self, peer_id: network::PeerId, best_block_number: u64) {
+    pub async fn raise_source_best_block(&self, peer_id: libp2p::PeerId, best_block_number: u64) {
         self.to_background
             .lock()
             .await
@@ -221,10 +221,10 @@ impl SyncService {
 
 /// Message sent to the background task.
 enum ToBackground {
-    PeerConnected(network::PeerId, u64),
-    PeerDisconnected(network::PeerId),
+    PeerConnected(libp2p::PeerId, u64),
+    PeerDisconnected(libp2p::PeerId),
     PeerRaiseBest {
-        peer_id: network::PeerId,
+        peer_id: libp2p::PeerId,
         best_block_number: u64,
     },
 }
@@ -234,7 +234,7 @@ enum ToBackground {
 enum FromBackground {
     /// A blocks request must be started.
     RequestStart {
-        target: network::PeerId,
+        target: libp2p::PeerId,
         request: network::protocol::BlocksRequestConfig,
         send_back: oneshot::Sender<Result<Vec<network::protocol::BlockData>, ()>>, // TODO: proper error
     },
@@ -253,7 +253,7 @@ async fn start_sync(
     mut from_foreground: mpsc::Receiver<ToBackground>,
     mut to_database: mpsc::Sender<ToDatabase>,
 ) -> impl Future<Output = ()> {
-    let mut sync = optimistic::OptimisticSync::<_, network::PeerId, ()>::new(optimistic::Config {
+    let mut sync = optimistic::OptimisticSync::<_, libp2p::PeerId, ()>::new(optimistic::Config {
         chain_information: database
             .to_chain_information(&database.finalized_block_hash().unwrap())
             .unwrap(),
