+++ src/network.rs
@@ -20,16 +20,8 @@
 **   Not ready yet                                      **
 *********************************************************/
 
-pub mod connection;
-pub mod discovery;
 pub mod leb128;
-pub mod libp2p;
-#[doc(inline)]
-pub use parity_multiaddr as multiaddr;
-pub mod peer_id;
-pub mod peerset;
 pub mod protocol;
 pub mod service;
 
 pub use parity_multiaddr::Multiaddr;
-pub use peer_id::PeerId;
